code_nth_cdr.feature added

+++ instructions/code/nth_cdr.rb
@@ -5,10 +5,12 @@ class CodeNthCdr < NudgeInstruction
   
   def process
     tree = NudgePoint.from(code(0))
+    raise NudgeError::InvalidScript,"code_nth_cdr cannot parse its argument" if tree.is_a?(NilPoint)
+    
     tree = BlockPoint.new(tree) unless tree.is_a?(BlockPoint)
     
     if tree.points > 1
-      n = [tree.points - 1, [int(0), 0].max].min
+      n = [tree.backbone_points, [int(0), 0].max].min
       n.times { tree.delete_point_at(1) }
     end
